LayoutTests/accessibility/mac/value-change-userinfo-expected.txt

0234567890
This tests value change notifications user info data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "0"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypeCut
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "1"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypeDelete
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "0"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypePaste
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "234567890"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypePaste
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "1"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypeTyping
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "1"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypeDelete
PASS successfullyParsed is true

TEST COMPLETE
